feat: create TypedError type
Creates a type that combines two error values using one as a value and
the other as a type. The purpose of the combination is to support
creating errors that belong to a category (the type) without requiring
the error representing the type to appear in the Unwrap chain and
without requiring custom implementations of the error interface.